feat(secret): new NixOS module for sharing secrets
Shared files are served via HTTP and encrypted with a preshared secret
key. Remote files are fetched only if they are not already present.

It is used to share Wireguard's public keys of the nodes with each
other. For more sensitive data, future work should allow for asymmetric
encryptions, the public keys being initially shared with this symmetric
scheme.

Options in the module are available to make it serve some files only
within the VPN. This must remain the default setting.