feat(pki): cfssl with shared root CA
The initialisation node creates a self-signed CA. Other nodes in the
cluster then fetch it with the secretSharing mechanism. CFSSL is started
on every node, so that Certmgr can be used to manage certificates.

Anyone knowing the shared secret key is able to fetch the CA and
impersonate nodes in the cluster. This should be fixed by strengthening
the secretSharing module at some point.