feat(kubernetes): use Wireguard for node-to-node traffic
The `vxlan` backend for Flannel seems to be broken. Instead, use
the CNI network plugin and configure Wireguard to route the packets
to the appropriate nodes.

The pod CIDR assigned to each node is now hard-coded in the Nix
configuration instead of being dynamically allocated by Kube's
controller manager, so we can set up Wireguard accordingly.